Three bindings say they are unused, and one conversion goes
Clippy runs with warnings as errors, so these would have failed the build.
Two tests were rewritten around helpers that need fewer of their fixture's
parts; the third lost the assertion that read its binding, and that assertion
now lives in `tests.ply`, which checks both no-tests messages.

+++ crates/ply-cli/src/test.rs
@@ -1124,7 +1124,7 @@ fn outcome(result: &TestResult) -> OutcomeView {
                 .map(|naive| (u64::from(naive.explored), naive.bounded, naive.to_string())),
             // Tenths, so one division answers both the line and the document.
             reduction_tenths: e.reduction().map(|r| (r * 10.0).round() as i64),
-            steps: u64::from(e.steps),
+            steps: e.steps,
             virtual_time_ns: e.virtual_time,
             failing_seed: e.failure.as_ref().map(|s| s.to_string()),
         }),
